refactor: 앱 구조를 이루는 파일(NavGraph 등)을 structure 패키지로 분리
